elasticsearch: Disable date_detection in origin mapping
So that if the first document with a given field has a date-like value
in that field, Elasticsearch does not enforce that all documents and
queries on that field use a data.

Also, in the same test checking this, add a similar check for numeric
values, in case 'numeric_detection' ever gets turned on by default.