Add basic support for an origin blocklist
This adds a boolean field, "blocklisted", to the origin documents. When
this field is true, the documents are not returned in search results.

The blocklist field is sticky for updates, meaning that another run of the
swh.search journal client (or further visits of the origin) will keep
the origin hidden from search results.