search: Ensure CodeMeta dates are properly formatted
Some date values that can be found in codemeta.json files (dateCreated,
dateModified, datePublished) might be in a format not parsable by
elasticsearch which prevents successfull update of origin intrinsic
metadata in elasticsearch indices.

For instance, the date 2021-7-23 cannot be parsed by elasticsearch
as it expects 2021-07-23 instead.

So ensure to properly format CodeMeta dates to avoid such indexing
errors.