conftest: Fix tests hang with elasticsearch 7.17.3
elasticsearch 7.17.3 uses JDK >= 18 which changed the value of
"java.security.manager" property to "disallow".

As a consequence a "java.lang.UnsupportedOperationException" is
now raised when bootstrapping elasticsearch server used in tests.

So explicitely set the "java.security.manager" property to "allow"
to avoid that exception raising when elasticsearch setups its custom
security manager.

Closes T4192