setup.py: Fix editable install with setuptools >= 64
The way editable installs are handled by setuptools has changed since
release >= 64 and the custom develop command in setup.py is no longer
called. As a consequence, the tree-sitter parser code generation was
no longer executed when performing an editable install and runtime
error was raised when executing the swh-search backend.

So use proper way to handle editable install with latest setuptools and
ensure backward compatibility with older versions.