query_language: Remove hardcoded list of origin visit types
There is no easy way to get that list dynamically as the query language
parser code in C is generated when building the swh-search Python package
and not at runtime.

So prefer to remove that harcoded list of visit types as it must updated
each time a new visit type is added in the archive or searching for such
visit type will fail.