interface: Update origin_search to return visits info of origins
Previously, origin_search was only returning a list of dict containing
a single url field. From now on, the list of visit types and a boolean
indicating if an origin was visited by SWH are also included in each
returned origin dict.

The purpose of that change is to enable the webapp to exploit the list
of visit types associated to an origin when searching for origins.

Related to swh/devel/swh-web#4786.