setup: fix editable install
Add missing links in editable install to parser source files generated
by tree-sitter.

Since the Translator can build the extension on the fly, this should
work ok for strict editable install.

Note that this also fixes the building of the sdist package, including
generated files to build the tree-sitter parser extension.

Thanks to lunar for this solution.