conftest.py: Ensure elasticsearch 8.x to use a reasonable amount of RAM
Since recent elasticsearch releases, if the jvm heap size is not explicitly
set es will use the half of the total amount of RAM as default heap size.

So on a laptop with 64Gb of RAM, a process eating 32Gb of RAM is spawned
when running swh-search tests.

Set a reasonable value of 512Mb instead.